// src/app/squarekles-game/enums/item-being-displayed.ts
export enum ItemBeingDisplayed {
  ReservedTopTierUnknownCard = 0,
  ReservedMiddleTierUnknownCard = 1,
  ReservedBottomTierUnknownCard = 2,
  ReservedTopTierCard0 = 3,
  ReservedTopTierCard1 = 4,
  ReservedTopTierCard2 = 5,
  ReservedTopTierCard3 = 6,
  ReservedMiddleTierCard0 = 7,
  ReservedMiddleTierCard1 = 8,
  ReservedMiddleTierCard2 = 9,
  ReservedMiddleTierCard3 = 10,
  ReservedBottomTierCard0 = 11,
  ReservedBottomTierCard1 = 12,
  ReservedBottomTierCard2 = 13,
  ReservedBottomTierCard3 = 14,
  PurchaseTopTierCard0 = 15,
  PurchaseTopTierCard1 = 16,
  PurchaseTopTierCard2 = 17,
  PurchaseTopTierCard3 = 18,
  PurchaseMiddleTierCard0 = 19,
  PurchaseMiddleTierCard1 = 20,
  PurchaseMiddleTierCard2 = 21,
  PurchaseMiddleTierCard3 = 22,
  PurchaseBottomTierCard0 = 23,
  PurchaseBottomTierCard1 = 24,
  PurchaseBottomTierCard2 = 25,
  PurchaseBottomTierCard3 = 26,
  TakeTokens = 27
}